In magic.t, restore $$ after modifying it.
Otherwise tests for setting $0 fail on any Linux or FreeBSD machine which
happens to have a running process 42. (The value $$ is set to for the
assignment test). Without a running process 42, those tests skip, on the
assumption that they aren't running on a suitable platform to read the process
name with external commands.