Remove Shell from the core distribution. Get it from CPAN now.
This reinstates 28d302d426b73ed7, which was reverted by c9a0cae924d6331f,
because Shell.pm was missing its deprecation warning in 5.12. (Minor changes
subsequently mean that c9a0cae924d6331f as-is can't be reverted).

The entry in perldelta.pod has to to be C<Shell> because L<Shell> would be a
broken link, and fail t/porting/podcheck.t. Change the removal notice for
Devel::DProf from L<> to C<> to be consistent. L<Devel::DProf> *isn't* a
podchecker error - presumably the "=head2 Devel::DProf" in perlperf.pod is seen
as the valid link target.