Benchmark.t: ignore sys CPU time
Currently we work out the CPU burned by a loop by summing user and sys
CPU. On the grounds that the burning we're interested in should have all
been carried out with user CPU and that therefore any sys CPU is a red
herring, ignore sys CPU for the infamous test 15.

Yes, this is clutching at straws. Still, diagnostics output may show
in future whether I was right!