charnames: Add :loose matching
This adds the capability to charnames to use Unicode loose name
look-ups, via ":loose" being specified in the pragma.

The number of tests per code point is doubled in the .t, so to preserve
the same amount of elapsed test time, the number of code points tested
in each run is halved.