CGI: Reset permissions during temporary directory test so Win32 can remove.
The 'tmpdir.t' test leaves the permissions of the directories at 0500, which
translates into setting the 'read-only' bit in Win32-speak.  That makes the
'rmdir' in the END block fail (silently) and thus the test is very unhappy
the next time through.  It is easy enough to just run the test twice to see.