Remove unnecessary code and tests
This test was relying on bug #91844.  As that has been fixed (by com-
mit 3ed94dc, which erroneously claimed to fix #81944), this test is
not testing what it purports to test.

Also as a result of that bug fix, it doesn’t matter if the TEMP flag
is set, I can get rid of the SvTEMP_off.  The only effects the TEMP
flag can have now occur only when the refcount is 1, but since *_
holds a refcount and the mortals stack does as well, it’s going to be
2 unless someone undefs *_, in which case the callback won’t have any
reference to it.  If it tries to make a weak reference, doing so will
turn TEMP off.