podcheck.t: Add --add_link option
L<> links in pod apply to anything that search cpan can find.  This
means that it applies to more things than podcheck can find in the Perl
distribution.  So, a list is kept in a data base of modules and man
pages that are known to exist, but aren't findable by podcheck.  This
list has to be maintained manually.  The new option makes adding to the
list easier.

The man page and instructions are somewhat tweaked as well.