podcheck: Make sure perltoc is known
perltoc is problematic because it is machine generated from constituent
pods.  This means that errors in them propgagate to it.  As a result,
commit e678c2947ab9ff776af461b393f3a3eecebab64a just skipped it.   But
this led to pods that link to it being flagged as having broken links.
This commit changes things so instead of skipping perltoc, it is added
to the list of pods that are known but are parsed only when there are
links to inside it (of which there are unlikely to be any), and then
only for the existence of such targets.