In Cwd::_win32_cwd() avoid a string eval when checking if we're miniperl.
To allow ExtUtils::MakeMaker to run tests as if it's "miniperl" we need to avoid
taking any sort of reference to the typeglob or the code in the optree, as its
test modules are loaded later than Cwd. Previously this was done with a string
eval, but that was causing problems (for unclear reasons - rt.cpan.org #56225).
Using a symbol table lookup and *foo{THING} syntax avoids the string eval.

Evolved from a suggestion by Reini Urban.