perldelta: Minor pod fixes
Commit a1e75797c204ade843c6eb8052cc5577af06c1d6 changed L<Shell>
to C<Shell>, but in reality there still is a Shell module accessible
from L<>, because L<> looks through all of CPAN.  podcheck.t has
to know that it's accessible, and that was done with commit
ce98560028e1830366c9182fe9dd70ec264fe6ee.

Also make the reference to Devel::DProf from C<> to L<> for symmetry.

And, a file reference should be enclosed in F<> and not C<>