use TestInit 'T' in porting{authors,cmp_version,manifest}.t to chdir '..'
These three tests run git commands from the top level of the build tree, so
need to chdir '..' if run from 't'. use TestInit 'T' to do this, as it simplify
adapting the tests to also work when the build tree is a symlink forest from
-Dmksymlinks. TestInit also sets $ENV{PERL5LIB}, avoiding the need to use -Ilib
on the command line of the programs they run.

t/porting/manifest.t now runs throughout at the top level, which simplifies it
slightly.