podcheck.t: sort --add_link output
The recently added --add_link option just appended lines to the end of
the data base file.  While this works, it means that the next time the
db is regen'd, that they will move around and show up in difference
listings unrelated to the original patch, thus potentially confusing
people.