No need for D and d flags in pod.lst - source is fixed, and contains the target
As the edited perldelta is now always pod/perldelta.pod (instead of a file with
version numbers in it), we don't need a 'D' flag to mark it. As perldelta.pod
contains within itself the number of the version it refers to, we can use that
to infer the 'd' flag for the target filename.

This simplifies the release manager's tasklist for version bumping.