Add perldelta "known issues" to the podchecker exception list.
Porting/perldelta_template.pod is exempted from t/porting/podchecker.t
Hence when it is copied to pod/perldelta.pod, exceptions need to be added for
certain template boilerplate it contains, which are not valid pod
(but ultimately will be edited out before the release).

This doesn't seem ideal, but I'm not sure of a better way to do this.