Stop ‘sub :lvalue :Foo’ from applying :lvalue to defined subs
This is something that commit 885ef6f missed.  See ticket #68758 and
the perldelta diff in this commit.

I thought that the prototype-application inconsistencies mention in
<2C7C1BF5-A814-4F5E-B10E-E431B71BA8CA@cpan.org> would get in the way
of this, but I found that the easiest way to fix it actually did not
change any of that.  So those issues can wait till another day.