The podchecker much prefers a forward slash in F<symbian/TODO>
It looks for the file within the F<> - on *nix, 'symbian\TODO' doesn't exist,
so the line is flagged as a known issue. On Windows, it does exist (because
at the API level, / and \ are both valid as directory separators), so there is
no known issue, and the test fails with a request to delete the known issue.

README.symbian is not consistent in its use of backward or forward slashes
inside F<> markup, so this isn't the first use of a forward slash.

(The podchecker has other exemption rules for F<>, hence why this doesn't
matter in some cases.)