Fix up some tests in sub_lval.t
One test was disabled long ago

Returning undef in list context was allowed by change 10777 (f206cdd)
in 2001.  There was already a test for disallowing it.  That test was
disabled by change 10779 (4c8a4e58), rather than correct to test the
new behaviour.

This commit changes it to test that undef return is allowed in list
context, also adding a test for explicit return.