Fix explicit return of pad var in list lv context
This is something that commit e08be60 missed, though it never worked
properly, even in 5.14, as explicit return from lvalue subs used to
copy return values.

As the commit message for e08be60 states, returning a scalar itself
from an lvalue sub does not work if it is a pad variable with a refer-
ence count of 1, because the sub-popping code clears it on exit.

The one code path that did not account for this was list lvalue con-
text (real lvalue context, not just potentially lvalue).  The only
observable effect this has is that assigning to a magic pad variable
returned from a subroutine in list context will not trigger set-magic.

This commit fixes it and also adds tests for returned magic pad vars
in all combinations of list/scalar lvalue/ref context.