op_lvalue .= _flags
Add flags param to op_lvalue, so that the caller can ask it not to
croak when encountering an unmodifiable op (upcoming).

This is in preparation for making the \$ prototype accept any lvalue.

There is no mathom, as the changes that this will support
are by no means suitable for maint.