Change dNOOP to a function declaration.
Previously dNOOP would declare an unused variable, and where possible mark it
as "unused" to prevent compiler warnings. However, clang doesn't provide
attribute unused, so will produced screenfuls of warnings if invoked with
-Wunused-variable, hiding the real unused variable warnings.

Using a function declaration instead avoids these warnings.

We can't just use define dNOOP as nothing, or a comment, as it is used as

    dNOOP;

and a bare semicolon is an error.