utf8_heavy.pl: Add INVERT_IT to swash
set a new key "INVERT_IT" in the swash object to whether or not
the file name for a property-value in Heavy.pl begins with a "!".
(No such entry currently exists.)  This will mean that the file's
contents need to be inverted to get the correct values for the
property-value.