mktables: Suppress complement tables
Consider the property \p{AHex}.  This is really \p{AHex=Y}, and there
is a corresponding \p{AHex=N} that is just the complement.  Prior to
this patch, 2 tables were generated for AHex.  Now, just one is, and the
N table is set-up in Heavy.pl to be the complement of the Y table.  This
saves quite a few tables and corresponding disk space.

Note that not all complements need be binary tables.