In cmp_version.t, show in the TAP when we are skipping a particular version.
The facility to skip specific versions of modules was added in
a5a2efce1e0ae6b7. However, it gave no indication when such an exception caused
a skip, because it generated identical TAP output to the normal pass case.

Instead, we now show the diffs (as for the "not ok" TAP output), but generate
TAP "SKIP" for the pass. This makes it clear what is happening.