mktables: Move test to added class member
Early Unicode releases didn't specify all possible property values,
but newer ones do, with some exceptions.  As an error check, on a newer
relese, a warning is raised if an unexpected property value is raised.
This member allows properties to be handled in new releases that
Unicode doesn't list the property values for.  (Each property value
corresponds to a mapping.)  The previous test that had a hard-coded
value is amended to use the member's value instead.