In sv.c, remove comments with structure sizes from bodies_by_type[].
On balance, the "typical" sizes seem to be more of a liability than an asset,
as they (by necessity) can only be correct for one particular combination of
UV size, pointer size, NV alignment, etc, which means more likely than not
they *aren't* correct for the platform you're building on. And that's for the
best case scenario, that the comments are updated if the structures are
updated.