No need for makedef.pl to invoke $^X -V to get 'Compile-time options'.
Config now exposes this as Config::bincompat_options() and
Config::non_bincompat_options().

Change WinCE to invoke makedef.pl with -MCross, which it should have been using
all along, as the values it wants are from the host config, not the target
config. [Not tested on WinCE]. Previously makedef.pl would have been loading
the target config, but as it was only using $Config{path_sep}, this would not
have mattered.