Bring cop label testing in line with intentions.
8375c93eec supplied tests for a newly exported API but the type
declarations for the arguments and return values didn't match the
types of the API being tested.

aebc0cbee0 renamed the functions without updating the calls to
those functions in the test.

Either of these could have been easily spotted by building with
g++ (or other readily available tools) before pushing.

This aside from the controversy over whether this particular API
is the one that should be publicly exported, so this all may be
reverted and replaced before 5.16.0, but hopefully it will unbreak
the build for now.