Make tests for gmagic on sub exit count fetches
These tests, added in commit 767eda44 (when lvalue subs started auto-
vivifying, causing regular sub exit to have to call get-magic), were
not counting the number of fetches, but simply whether the fetches
were happening.  The number that was occurring was wrong until
7ffa7e75, which fixed symbolic refs to call magic the right number of
times.  But now that that’s fixed, we can count them.  So this commit
does just that.