Call get-magic on temp returned under recursion
This fixes a regression introduced in 5.15.0.

Commit 767eda44 made autovivification work with scalars returned from
lvalue subs.  In doing so, as it made entersub an autovivifying op,
it had to add SvGETMAGIC calls to pp_return and pp_leavesub.  (See
767eda44’s commit message for details.)  In one spot in pp_return,
the SvGETMAGIC call was added to the wrong branch of an if/else con-
dition, causing a regression.  Explicitly returning from a recursive
subroutine call in autovivifying context would not call get-magic on
the returned value.