[perl #95538] perldiag.pod on defined %hash
I found the perldiag.pod description of item "defined(%hash) is
deprecated" unclear.

At first I thought it was threatening another incompatible change, but
if I'm not mistaken it's merely that defined %hash has only ever been a
check for non-empty, which may not be what you intended.  If so then
perhaps a wording like the following could make it clearer,