Add core_prototype; make pp_prototype use it
This commit moves the code for generating core prototypes into a sepa-
rate function, core_prototype, in op.c. This serves two porpoises:

• It allows the lock and tie exceptional cases to be incorporated into
  the main prototype=generation code, which requires the use of a
  static function in op.c.
• It allows other parts of the core (e.g., the upcoming \&CORE::foo
  feature) to use the same code.

The docs for it are in a section boringly entitled ‘Functions in
op.c’, for lack of a better name. This, I believe, is the only op.c
function that is in perlintern currently, so it’s hard to see what to
name a section that will, at least for now, contain nothing else.