core_prototype: Eliminate the special mkdir case
The prototype-generation code just needed a little tweaking
for this to work.  It assumed that ‘;’ should not be emitted if
PL_opargs[opnum] & OA_DEFGV, which is not necessarily the case.  It
only applies at the beginning of the prototype, hence the n check
(where n is the character position in the protoytpe).  It also changed
the last $ to _ in the OA_DEFGV case, instead of the first.  _ only
comes at the beginning of a prototype (at least for core functions;
ck_fun, incidentally, enforces this), but not necessarily at the end.