In makedef.pl, change readvar() to add items to a passed-in hash reference.
Previously, readvar() returned a list of symbols, extracting the names of
the variables declared there, then transforming them either via a passed in
function, or prepending 'PL_'. The list was then either added to %skip, or
added to %export by try_symbols(), which had a kludge to normalise the
previous transformation to 'PL_'.

Instead, passing a reference to the target hash into readvar() allows that
kludge to be removed, considerably simplifying try_symbols().