In makedef.pl, tidy up the USE_PERLIO logic.
No need to call try_symbols() for PL_perlio_mutex, as it is found in
perlvars.h, and no need to duplicate the skip for it if USE_ITHREADS is not
defined.

We do need to skip PL_perlio_mutex if if USE_PERLIO is not defined but
USE_ITHREADS is [not that this is a configuration that Configure will let
you select, until you edit it :-)]

Add a comment describing what the "Oddities from PerlIO" are.

Remove PerlIO_sv_dup from the list, as it's also in @layer_syms, and
inclusion in @layer_syms makes the former redundant. It was added to both in
commit 8437356be6fd345c in 2001.