podcheck.t: Move perldelta placeholder link checks
podcheck has the capability to ignore certain broken links in
perldelta.pod.  This is because this pod is special, initialized to a
template with various placeholder text including links that are to
eventually be changed to their correct values.  One of the final steps
in making a release is to clean it up, removing any of these links that
haven't been changed, and hence don't apply to the current release.

To lessen the number of steps in making a release, a list of these links is
hard-coded into podcheck, and it doesn't warn on those.  The check to
skip these links prior to this commit did not easily allow links of the
form page/node to be checked for.