release managers guide: perldelta broken link handling
podcheck.t contains a list of placeholder links in perldelta that don't
point to a real target, and hence shouldn't generate messages.  This
list is to make the release manager's job easier.  But if new
placeholder links are created, it may be that they have to be added to
the list.